코멘트 컴포넌트 나누기 함
