+++ .gitignore
@@ -90,3 +90,4 @@ fabric.properties
 !.idea/runConfigurations
 
 # End of https://www.toptal.com/developers/gitignore/api/intellij+all
+frontend_steam_game_review/db.json
