+++ frontend_steam_game_review/src/reduex/config/configToolket.js
@@ -1,9 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
 import loginSignUp from "../modules/loginSignUp";
 import comments from "../modules/commentsSlice";
+import games from "../modules/gameInfoSlice";
 
 const store = configureStore({
-  reducer: { loginSignUp: loginSignUp, comments },
+  reducer: { loginSignUp: loginSignUp, comments, games },
 });
 
 export default store;
